# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Website Socav RDC API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/createAbout": {
            "post": {
                "tags": [
                    "About"
                ],
                "summary": "Create",
                "description": "Creation",
                "operationId": "storeAbout",
                "requestBody": {
                    "description": "Enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "objective_fr",
                                    "objective_en",
                                    "title_en",
                                    "title_fr",
                                    "short_description_en",
                                    "short_description_fr",
                                    "adress_en",
                                    "adress_fr",
                                    "facebook",
                                    "instagram",
                                    "linkedin",
                                    "twitter",
                                    "email",
                                    "phone",
                                    "youtube",
                                    "image1",
                                    "image2"
                                ],
                                "properties": {
                                    "objective_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "objective_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "short_description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "short_description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "adress_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "adress_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "linkedin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "youtube": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateAbout/{id}": {
            "post": {
                "tags": [
                    "About"
                ],
                "summary": "Update",
                "description": "Update existing About entry",
                "operationId": "updateAbout",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the About entry to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update About information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "objective_fr",
                                    "objective_en",
                                    "title_en",
                                    "title_fr",
                                    "short_description_en",
                                    "short_description_fr",
                                    "adress_en",
                                    "adress_fr",
                                    "facebook",
                                    "instagram",
                                    "linkedin",
                                    "twitter",
                                    "email",
                                    "phone",
                                    "youtube",
                                    "image1",
                                    "image2"
                                ],
                                "properties": {
                                    "objective_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "objective_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "short_description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "short_description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "adress_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "adress_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "linkedin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner@gmail.com"
                                    },
                                    "youtube": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mpesa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deleteAbout/{id}": {
            "delete": {
                "tags": [
                    "About"
                ],
                "summary": "Delete",
                "description": "Delete existing About entry",
                "operationId": "deleteAbout",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the About entry to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}